==> Part2.Rmd <==
---
title: "Habitat complexity part 2: visualisation of benthic cover"
author: "Alberto Rovellini"
date: "18 March 2020"
output:
  html_document:
    code_folding: hide
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
---

This document visualises the main properties of the 32 quadrats. 

```{r, echo = FALSE, include=FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = TRUE, width = 150)
```

```{r, message = FALSE, warning = FALSE}

library(readxl)
library(data.table)
library(ggplot2)
library(abind)
library(dplyr)
library(reshape2)
library(RColorBrewer)
library(ggfortify)

```

Read in cover data.

```{r}

cover <- read.csv(".../coverNew.csv")

```

```{r}

# ditch morphs for generic benthic cover

no.morphs <- cover %>% dplyr::group_by(
  Site, Quadrat, Type, Complexity, Group, Label
) %>% dplyr::summarise(
  Cover = sum(Percent)
)

# change labels

no.morphs <- no.morphs %>% dplyr::mutate(
  Label = substr(Label, 4, 6)
)

```

Plot benthic composition of the 32 quadrats as stacked bars.

```{r, fig.width = 7, fig.height = 4.5}

# prepare palette

col.count <- length(unique(no.morphs$Group))
get.palette <- colorRampPalette(brewer.pal(14, "Spectral"))



p <- ggplot(data = no.morphs, aes(x = Label, y = Cover, group = Group, fill = Group))+
  geom_bar(stat = "identity", position = "stack", color = "white")+
  scale_fill_manual(values = get.palette(col.count))+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 45, hjust = 1))+
  theme(panel.grid = element_blank())
p

```

The above shows that we have 15 quadrats where hard coral is the single most abundant benthic category, and 17 quadrats where sponges are the single most abundant benthic category. This supports our definition of "coral-dominated" versus "sponge-dominated" quadrats.

List all the values of % cover of the dominant groups and see the lowest.

```{r}

coral.dominated <- no.morphs[no.morphs$Type == "Coral" & no.morphs$Group == "LC",] 

coral.dominated %>% 
  arrange(desc(Cover)) %>%
  ungroup() %>%
  select(Quadrat, Cover)

c(mean(coral.dominated$Cover), sd(coral.dominated$Cover)/sqrt(length(coral.dominated$Cover)))

```

```{r}

sponge.dominated <- no.morphs[no.morphs$Type == "Sponge" & no.morphs$Group == "SG",]

sponge.dominated %>% 
  arrange(desc(Cover)) %>%
  ungroup() %>%
  select(Quadrat, Cover)

c(mean(sponge.dominated$Cover), sd(sponge.dominated$Cover)/sqrt(length(sponge.dominated$Cover)))

```

Now see the proportion that the morphs make up of coral and sponges. Ditch all other benthic groups.

```{r, fig.width = 12, fig.height = 7}

cover.cs <- cover[cover$Group %in% c("LC", "SG"),] # keep only corals and sponges

# reorder factor for complexity

cover.cs$Complexity <- factor(cover.cs$Complexity, levels = c("L", "H"))

# reorder morphs so they go from high to low complexity (for visualisation - coloring purposes)

cover.cs$Morphology <- factor(cover.cs$Morphology, levels = c("BRA", "TUB", "TAB", "REP", "FOL", "DIG", "BAR", "MAS", "SUB", "CUS", "GLO", "FRL", "ENC"))

# prepare palette

col.count <- length(unique(cover.cs$Morphology))
get.palette <- colorRampPalette(brewer.pal(11, "Spectral"))

```

Get percentage cover of specific morphologies of corals and sponges in quadrats of different complexity and benthic type

```{r}

cover.cs %>% dplyr::group_by(
  Complexity, Type, Long_group
) %>% summarise(
  Mean = mean(Percent)
)

```


```{r, fig.width = 6, fig.height = 3.5}

col.count <- length(unique(cover.cs[cover.cs$Type == "Coral" & cover.cs$Group == "LC",]$Morphology))
get.palette <- colorRampPalette(brewer.pal(7, "RdYlBu"))

p.morphs.coral <- ggplot(data = cover.cs[cover.cs$Type == "Coral" & cover.cs$Group == "LC",], aes(x = Label, y = Percent, Group = Morphology, fill = Morphology))+
  geom_bar(stat = "identity", position = "stack", color = "grey70")+
  scale_fill_manual(values = get.palette(col.count))+
  scale_y_continuous(breaks = seq(0, 90, 10), limits = c(0, 90))+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 45, hjust = 1))+
  theme(panel.grid = element_blank())
p.morphs.coral

```

```{r, fig.width = 6, fig.height = 3.5}

col.count <- length(unique(cover.cs[cover.cs$Type == "Sponge" & cover.cs$Group == "SG",]$Morphology))
get.palette <- colorRampPalette(brewer.pal(10, "RdYlBu"))

p.morphs.sponges <- ggplot(data = cover.cs[cover.cs$Type == "Sponge" & cover.cs$Group == "SG",], aes(x = Label, y = Percent, Group = Morphology, fill = Morphology))+
  geom_bar(stat = "identity", position = "stack", color = "grey70")+
  scale_fill_manual(values = get.palette(col.count))+
  scale_y_continuous(breaks = seq(0, 90, 10), limits = c(0, 90))+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 45, hjust = 1))+
  theme(panel.grid = element_blank())
p.morphs.sponges

```

Mean cover of all morphologies for each level of complexity and dominant benthic organism. 

```{r}

mean.cover <- cover.cs %>% dplyr::group_by(
  Type, Complexity, Long_group
) %>% dplyr::summarize(
  Mean = mean(Percent),
  SE   = sd(Percent)/length(Percent)
)

mean.cover

```
